backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
drf-spectacular>=0.28,<0.29
django-cors-headers>=4.6,<5.0
Pillow>=10.0,<11.0

python-dotenv

gunicorn
